better organised code
