+++ pytorch_lightning/utilities/memory.py
@@ -1,3 +1,6 @@
+import gc
+import torch
+
 def recursive_detach(in_dict: dict) -> dict:
     """Detach all tensors in `in_dict`.
 
@@ -20,3 +23,38 @@ def recursive_detach(in_dict: dict) -> dict:
         else:
             out_dict.update({k: v})
     return out_dict
+
+
+def is_OOM_error(exception):
+    return is_cuda_out_of_memory(exception) or is_cudnn_snafu(exception) or is_out_of_cpu_memory(exception)
+
+
+def is_cuda_out_of_memory(exception):
+    return (
+        isinstance(exception, RuntimeError) and
+        len(exception.args) == 1 and
+        "CUDA out of memory." in exception.args[0]
+    )
+
+
+def is_cudnn_snafu(exception):
+    return (
+        isinstance(exception, RuntimeError) and
+        len(exception.args) == 1 and
+        "cuDNN error: CUDNN_STATUS_NOT_SUPPORTED." in exception.args[0]
+    )
+
+
+def is_out_of_cpu_memory(exception):
+    return (
+        isinstance(exception, RuntimeError) and
+        len(exception.args) == 1 and
+        "DefaultCPUAllocator: can't allocate memory" in exception.args[0]
+    )
+
+
+def garbage_collection_cuda():
+    """Garbage collection Torch (CUDA) memory."""
+    gc.collect()
+    if torch.cuda.is_available():
+        torch.cuda.empty_cache()
